Save log only if large enough

+++ mainwindow.cpp
@@ -66,11 +66,13 @@ MainWindow::MainWindow(QWidget *parent)
 MainWindow::~MainWindow()
 {
     QString tmpLog = "/tmp/" + QApplication::applicationName() + ".log";
-    QString log = "/var/log/" + QApplication::applicationName() + ".log";
-    if (QFileInfo::exists(log)) {
-        shell->runAsRoot("mv --backup=numbered " + log + " " + log + ".old", nullptr, nullptr, true);
+    if (Cmd().getCmdOut("wc -l " + tmpLog + "| cut -f1 -d' '").toInt() > 5) { // only if log is large enough
+        QString log = "/var/log/" + QApplication::applicationName() + ".log";
+        if (QFileInfo::exists(log)) {
+            shell->runAsRoot("mv --backup=numbered " + log + " " + log + ".old", nullptr, nullptr, true);
+        }
+        shell->runAsRoot("cp " + tmpLog + " " + log, nullptr, nullptr, true);
     }
-    shell->runAsRoot("cp " + tmpLog + " " + log, nullptr, nullptr, true);
     delete ui;
 }
